Refactor operations
 - Move annotation matchers and mergers to the seperate directory
 - Move get_merger to the seperate directory
