+++ .pre-commit-config.yaml
@@ -61,4 +61,3 @@ repos:
     hooks:
       - id: nbqa-black
       - id: nbqa-ruff
-        language_version: python3.8
