+++ datumaro/components/annotations/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (C) 2023 Intel Corporation
+#
+# SPDX-License-Identifier: MIT
+
+from .matcher import *
+from .merger import *
